+++ htof/test/test_main.py
@@ -39,7 +39,7 @@ def test_parse_and_fit_to_line():
 
 
 @pytest.mark.e2e
-def test_Hip2_fit_to_known_system():
+def test_Hip2_fit_to_hip27321():
     # Hip 27321 parameters from Snellen+Brown 2018: https://arxiv.org/pdf/1808.06257.pdf
     cntr_ra, cntr_dec = Angle(86.82118054, 'degree'), Angle(-51.06671341, 'degree')
     plx = 51.44  # mas
@@ -48,7 +48,7 @@ def test_Hip2_fit_to_known_system():
     # generate fitter and parse intermediate data
     astro = Astrometry('Hip2', '27321', 'htof/test/data_for_tests/Hip2', central_epoch_ra=1991.25,
                        central_epoch_dec=1991.25, format='jyear', fit_degree=1, use_parallax=True,
-                       central_ra=cntr_ra, central_dec=cntr_dec)
+                       central_ra=cntr_ra, central_dec=cntr_dec, normed=False)
     chisq = np.sum(astro.data.residuals ** 2 / astro.data.along_scan_errs ** 2)
     # generate ra and dec for each observation.
     year_epochs = Time(astro.data.julian_day_epoch(), format='jd', scale='tcb').jyear - \
@@ -63,11 +63,43 @@ def test_Hip2_fit_to_known_system():
     #
     coeffs, errors, chisq_found = astro.fit(ra.mas, dec.mas, return_all=True)
     assert np.isclose(chisq, chisq_found, atol=1E-3)
-    assert np.allclose([plx, pmRA, pmDec], np.array([coeffs[0], coeffs[3], coeffs[4]]).round(2))
+    assert np.allclose([pmRA, pmDec], np.array([coeffs[3], coeffs[4]]).round(2))
+    assert np.isclose(plx, coeffs[0].round(2), atol=0.01)
+    assert np.allclose(errors.round(2), np.array([0.11, 0.10, 0.11, 0.11, 0.15]))
 
 
 @pytest.mark.e2e
-def test_Hip1_fit_to_known_system():
+def test_hip2_fit_to_hip78999():
+    cntr_ra, cntr_dec = Angle(241.89259265, 'degree'), Angle(-5.70677966, 'degree')
+    plx = 28.11  # mas
+    pmRA = 156.38  # mas/year
+    pmDec = -177.64  # mas/year
+    # generate fitter and parse intermediate data
+    astro = Astrometry('Hip2', '78999', 'htof/test/data_for_tests/Hip2', central_epoch_ra=1991.25,
+                       central_epoch_dec=1991.25, format='jyear', fit_degree=1, use_parallax=True,
+                       central_ra=cntr_ra, central_dec=cntr_dec, normed=False)
+    chisq = np.sum(astro.data.residuals ** 2 / astro.data.along_scan_errs ** 2)
+    # generate ra and dec for each observation.
+    year_epochs = Time(astro.data.julian_day_epoch(), format='jd', scale='tcb').jyear - \
+                  Time(1991.25, format='decimalyear').jyear
+    ra_motion = astro.fitter.parallactic_pertubations['ra_plx']
+    dec_motion = astro.fitter.parallactic_pertubations['dec_plx']
+    ra = Angle(ra_motion * plx + pmRA * year_epochs, unit='mas')
+    dec = Angle(dec_motion * plx + pmDec * year_epochs, unit='mas')
+    # add residuals
+    ra += Angle(astro.data.residuals.values * np.sin(astro.data.scan_angle.values), unit='mas')
+    dec += Angle(astro.data.residuals.values * np.cos(astro.data.scan_angle.values), unit='mas')
+    #
+    coeffs, errors, chisq_found = astro.fit(ra.mas, dec.mas, return_all=True)
+    assert np.isclose(chisq, chisq_found, atol=1E-3)
+    assert np.allclose([pmRA, pmDec], np.array([coeffs[3], coeffs[4]]).round(2), atol=0.01)
+    assert np.isclose(plx, coeffs[0].round(2), atol=0.01)
+    # testing the errors. Note these values must come from the CD if we are testing IAD from the CD.
+    assert np.allclose(errors.round(2), np.array([2.4, 1.79, 0.94, 4.05, 2.2]))
+
+
+@pytest.mark.e2e
+def test_Hip1_fit_to_hip27321():
     # Hip 27321 parameters from the Hipparcos 1 catalogue via Vizier
     cntr_ra, cntr_dec = Angle(86.82118054, 'degree'), Angle(-51.06671341, 'degree')
     plx = 51.87  # mas
@@ -76,7 +108,7 @@ def test_Hip1_fit_to_known_system():
     # generate fitter and parse intermediate data
     astro = Astrometry('Hip1', '27321', 'htof/test/data_for_tests/Hip1', central_epoch_ra=1991.25,
                        central_epoch_dec=1991.25, format='jyear', fit_degree=1, use_parallax=True,
-                       central_ra=cntr_ra, central_dec=cntr_dec)
+                       central_ra=cntr_ra, central_dec=cntr_dec, normed=False)
     chisq = np.sum(astro.data.residuals ** 2 / astro.data.along_scan_errs ** 2)
     # generate ra and dec for each observation.
     year_epochs = Time(astro.data.julian_day_epoch(), format='jd', scale='tcb').jyear - \
@@ -91,25 +123,31 @@ def test_Hip1_fit_to_known_system():
     #
     coeffs, errors, chisq_found = astro.fit(ra.mas, dec.mas, return_all=True)
     assert np.isclose(chisq, chisq_found, atol=1E-3)
-    assert np.allclose([plx, pmRA, pmDec], np.array([coeffs[0], coeffs[3], coeffs[4]]).round(2))
+    assert np.allclose([pmRA, pmDec], np.array([coeffs[3], coeffs[4]]).round(2))
+    assert np.isclose(plx, coeffs[0].round(2), atol=0.01)
+    assert np.allclose(errors.round(2), np.array([0.51, 0.45, 0.46, 0.53, 0.61]))
 
 
-class TestAstrometry:
-    @pytest.mark.e2e
-    def test_astrometric_fit(self):
-        """
-        Tests fitting a line to fake RA and DEC data which has errors calculated from the real intermediate data
-        from Hip1, Hip2, and GaiaDR2.
-        """
-        stars = ['049699', '027321', '027321']
-        data_choices = ['GaiaDR2', 'Hip1', 'Hip2']
-        base_directory = os.path.join(os.getcwd(), 'htof/test/data_for_tests')
-        for star_id, data_choice in zip(stars, data_choices):
-            test_data_directory = os.path.join(base_directory, data_choice)
-
-            fitter = Astrometry(data_choice, star_id, test_data_directory, central_epoch_ra=2000,
-                                central_epoch_dec=2001, format='decimalyear')
-            num_pts = len(fitter.data.julian_day_epoch())
-            ra0, dec0, mu_ra, mu_dec = fitter.fit(ra_vs_epoch=np.ones(num_pts),
-                                                  dec_vs_epoch=np.ones(num_pts))
-            assert True
+@pytest.mark.e2e
+def test_Hip1_fit_to_hip27321_no_parallax():
+    # Hip 27321 parameters from the Hipparcos 1 catalogue via Vizier
+    pmRA = 4.65  # mas/year
+    pmDec = 81.96  # mas/year
+    # generate fitter and parse intermediate data
+    astro = Astrometry('Hip1', '27321', 'htof/test/data_for_tests/Hip1', central_epoch_ra=1991.25,
+                       central_epoch_dec=1991.25, format='jyear', fit_degree=1,
+                       use_parallax=False, normed=False)
+    chisq = np.sum(astro.data.residuals ** 2 / astro.data.along_scan_errs ** 2)
+    # generate ra and dec for each observation.
+    year_epochs = Time(astro.data.julian_day_epoch(), format='jd', scale='tcb').jyear - \
+                  Time(1991.25, format='decimalyear').jyear
+    ra = Angle(pmRA * year_epochs, unit='mas')
+    dec = Angle(pmDec * year_epochs, unit='mas')
+    # add residuals
+    ra += Angle(astro.data.residuals.values * np.sin(astro.data.scan_angle.values), unit='mas')
+    dec += Angle(astro.data.residuals.values * np.cos(astro.data.scan_angle.values), unit='mas')
+    #
+    coeffs, errors, chisq_found = astro.fit(ra.mas, dec.mas, return_all=True)
+    assert np.isclose(chisq, chisq_found, atol=1E-3)
+    assert np.allclose([pmRA, pmDec], np.array([coeffs[2], coeffs[3]]).round(2))
+    assert np.allclose(errors.round(2), np.array([0.45, 0.46, 0.53, 0.61]), atol=0.01)